Make game.prev() always go back one ply
When using prev() from the beginning of a variation,
game.ascenFromCurrentContinuation() was called so far. That means that
the cursor ends up two ply before the old position. So one call of
prev() would cause a shift by two ply. This behaviour seems
counterintuitive.

prev() uses ascendFromCurrentVariation() now which means that it will
always cause a shift of one ply back.
The wildcard move test in test/chess.js had to be adjusted to this new
behaviour.